Some annotation stuff

+++ core/data/src/main/java/g/sig/questweaver/data/dto/ColorDto.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.Serializable
 
 @JvmInline
 @Serializable
-value class ColorDto(val value: UInt) : Dto
+value class ColorDto(val value: ULong) : Dto
